Play around with more configs
