+++ examples/example1/.gitignore
@@ -1,1 +1,2 @@
 .next
+dist
