+++ examples/example1/.oxlintrc.json
@@ -1,0 +1,5 @@
+{
+  "rules": {
+    "no-alert": "error"
+  }
+}
