+++ examples/example1/apps/spa/index.html
@@ -1,1 +1,1 @@
-<script type="module" src="./index.js"></script>
+<script type="module" src="./src/index.js"></script>
