--- examples/example1/apps/spa/index.ts
@@ -1,3 +1,0 @@
-import console from '@repo/logger';
-
-console.info('SPA');
