+++ examples/example1/apps/spa/src/index.ts
@@ -1,0 +1,3 @@
+import { log } from "@repo/logger";
+
+log("SPA");
