+++ examples/example1/apps/spa/vite.config.ts
@@ -1,0 +1,5 @@
+import { defineConfig } from "vite-plus";
+
+export default defineConfig({
+  base: "web"
+});
