--- examples/example1/apps/spa/viteplus.config.ts
@@ -1,5 +1,0 @@
-import { defineConfig } from 'vite-plus';
-
-// automatically merge with root config? unless explicitly opted out
-
-export default defineConfig();
