+++ examples/example1/packages/logger/.oxlintrc.json
@@ -1,0 +1,6 @@
+{
+  "extends": ["../../.oxlintrc.json"],
+  "rules": {
+    "no-eval": "error"
+  }
+}
