+++ examples/example1/packages/logger/vite.config.ts
@@ -1,0 +1,14 @@
+import { join } from "node:path";
+import { defineConfig } from "vite-plus";
+
+export default defineConfig({
+  extends: "../../vite.config.ts",
+  resolve: {
+    alias: {
+      "@log": `${join(import.meta.dirname, "src")}`
+    }
+  },
+  test: {
+    reporters: ["dot"]
+  }
+});
