+++ examples/example1/scripts/script1.ts
@@ -1,0 +1,1 @@
+export const script1 = () => "script 1";
