+++ examples/example1/vite.config.ts
@@ -1,0 +1,10 @@
+import { join } from "node:path";
+import { defineConfig } from "vite-plus";
+
+export default defineConfig({
+  resolve: {
+    alias: {
+      "@scripts": join(import.meta.dirname, "scripts")
+    }
+  }
+});
