--- examples/example1/viteplus.config.ts
@@ -1,9 +1,0 @@
-import { defineConfig } from 'vite-plus';
-
-export default defineConfig({
-  resolve: {
-    alias: {}
-  },
-  dev: {},
-  build: {}
-});
