+++ packages/cli/package.json
@@ -24,12 +24,13 @@
   ],
   "dependencies": {
     "@oxc-node/core": "^0.0.27",
+    "@vitest/utils": "^3.2.4",
     "multiplexer": "workspace:*",
     "oxlint": "^0.16.5",
     "tinyexec": "^1.0.1",
     "tsdown": "^0.12.1",
     "vite": "npm:rolldown-vite@6.3.12",
     "vitepress": "^2.0.0-alpha.5",
-    "vitest": "^3.1.4"
+    "vitest": "^3.2.4"
   }
 }
